pyproject.toml
[project]
name = "opus-blocks"
version = "0.1.0"
description = "Grounded scientific writing assistant backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "bcrypt>=4.1.3",
    "python-multipart>=0.0.9",
    "celery>=5.3.6",
    "redis>=5.0.4",
    "openai>=1.30.0",
    "email-validator>=2.1.1",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=5.0.0",
]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.4",
    "mypy>=1.10.0",
    "pre-commit>=3.7.1",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-q"]
